+++ esmerald/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "3.7.5"
+__version__ = "3.7.6"
 
 from typing import TYPE_CHECKING